+++ package.json
@@ -12,27 +12,28 @@
     "dev:api": "wait-on api/index.js && nodemon api/index.js"
   },
   "dependencies": {
-    "@fastify/cors": "^7.0.0",
+    "@fastify/cors": "^8.0.0",
     "cli-color": "^2.0.2",
-    "cli-progress": "^3.11.0",
-    "fastify": "^3.29.0",
+    "cli-progress": "^3.11.1",
+    "fastify": "^4.0.2",
+    "fastify-plugin": "^3.0.1",
     "node-emoji": "^1.11.0",
     "pg": "^8.7.3",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
     "@types/cli-color": "^2.0.2",
-    "@types/cli-progress": "^3.9.2",
+    "@types/cli-progress": "^3.11.0",
     "@types/keyv": "^3.1.4",
-    "@types/node": "^17.0.31",
+    "@types/node": "^17.0.43",
     "@types/node-emoji": "^1.8.1",
     "@types/pg": "^8.6.5",
-    "concurrently": "^7.1.0",
+    "concurrently": "^7.2.2",
     "copyfiles": "^2.4.1",
-    "dotenv": "^16.0.0",
+    "dotenv": "^16.0.1",
     "nodemon": "^2.0.16",
     "rimraf": "^3.0.2",
-    "typescript": "^4.6.4",
+    "typescript": "^4.7.3",
     "wait-on": "^6.0.1"
   },
   "nodemonConfig": {
