+++ src/modulos/errata.ts
@@ -1,7 +1,6 @@
 export type Tablas = {
   [key: string]: {
     sinPorcentaje: number;
-    errores: string[];
     procesados: number;
     numeradorMayorQueDenominador: number;
     total: number;
