+++ src/rutas/indicadores.ts
@@ -1,0 +1,5 @@
+// import { FastifyInstance } from "fastify";
+
+// export default (servidor: FastifyInstance) => {
+//   servidor.get('/')
+// }
